Chanakya_Shield/requirements.txt
numpy
pandas
scikit-learn
joblib
scapy
mysql-connector-python
requests
dwave-neal
python-dotenv
lightgbm
imbalanced-learn
